// src/html/js/types/web.d.ts
interface HIDConnectionEventInit {
    device: HIDDevice;
}

/*~ https://wicg.github.io/webhid/#hidconnectionevent-interface */
declare class HIDConnectionEvent extends Event {
    constructor(type: string, eventInitDict: HIDConnectionEventInit);

    readonly device: HIDDevice;
}

/*~ https://wicg.github.io/webhid/#hidinputreporteventinit-dictionary */
interface HIDInputReportEventInit extends EventInit {
    device: HIDDevice;
    reportId: number;
    data: DataView;
}

/*~ https://wicg.github.io/webhid/#hidinputreportevent-interface */
declare class HIDInputReportEvent extends Event {
    constructor(type: string, eventInitDict: HIDInputReportEventInit);

    readonly device: HIDDevice;
    readonly reportId: number;
    readonly data: DataView;
}
// ...
interface HIDReportInfo {
    reportId?: number | undefined;
    items?: HIDReportItem[] | undefined;
}

/*~ https://wicg.github.io/webhid/#hidcollectioninfo-dictionary */
interface HIDCollectionInfo {
    usagePage?: number | undefined;
    usage?: number | undefined;
    type?: number | undefined;
    children?: HIDCollectionInfo[] | undefined;
    inputReports?: HIDReportInfo[] | undefined;
    outputReports?: HIDReportInfo[] | undefined;
    featureReports?: HIDReportInfo[] | undefined;
}

/*~ https://wicg.github.io/webhid/#hiddevice-interface */
declare class HIDDevice extends EventTarget {
    oninputreport: ((this: this, ev: HIDInputReportEvent) => any) | null;
    readonly opened: boolean;
    readonly vendorId: number;
    readonly productId: number;
    readonly productName: string;
    readonly collections: HIDCollectionInfo[];

    open(): Promise<void>;

    close(): Promise<void>;

    forget(): Promise<void>;

    sendReport(reportId: number, data: BufferSource): Promise<void>;

    sendFeatureReport(reportId: number, data: BufferSource): Promise<void>;

    receiveFeatureReport(reportId: number): Promise<DataView>;

    addEventListener(
        type: "inputreport",
        listener: (this: this, ev: HIDInputReportEvent) => any,
        options?: boolean | AddEventListenerOptions,
    ): void;
    addEventListener(
        type: string,
        listener: EventListenerOrEventListenerObject | null,
        options?: boolean | AddEventListenerOptions,
    ): void;

    removeEventListener(
        type: "inputreport",
        callback: (this: this, ev: HIDInputReportEvent) => any,
        options?: boolean | AddEventListenerOptions,
    ): void;
    removeEventListener(
        type: string,
        callback: EventListenerOrEventListenerObject | null,
        options?: EventListenerOptions | boolean,
    ): void;
}